+++ src/realm/objc/RLMTable.h
@@ -55,6 +55,8 @@
 // Getting and setting individual rows with object subscripting
 -(id)objectAtIndexedSubscript:(NSUInteger)rowIndex;
 -(void)setObject:(id)newValue atIndexedSubscript:(NSUInteger)rowIndex;
+-(RLMRow *)objectForKeyedSubscript:(NSString *)key;
+-(void)setObject:(id)newValue forKeyedSubscript:(NSString *)key;
 
 // Add a row at the end of the table.
 // If data is nil, an empty row with default values is added.
@@ -62,6 +64,9 @@
 
 // Inserting rows at specific positions
 -(void)insertRow:(NSObject *)anObject atIndex:(NSUInteger)rowIndex;
+
+// Updating rows at specific positions
+-(void)updateRow:(NSObject *)anObject atIndex:(NSUInteger)rowIndex;
 
 // Removing rows
 -(void)removeAllRows;
@@ -81,9 +86,6 @@
 // Indexing
 -(void)createIndexInColumnWithIndex:(NSUInteger)colIndex;
 -(BOOL)isIndexCreatedInColumnWithIndex:(NSUInteger)colIndex;
-
-// Optimizing
--(BOOL)optimize;
 
 // Table type and schema
 -(BOOL)isEqual:(id)otherTableClass;