+++ release_notes.md
@@ -39,10 +39,13 @@ The Objective-C API has been updated and your code will break!
 ### API breaking changes:
 
 * Changed all class prefixes from TDB to RLM
+* Write transactions now take an optional rollback parameter (rather than needing to return a boolean)
 
 ### Enhancements:
 
-* `createTableWithName:columns:` has been added to `TDBTransaction`.
+* `createTableWithName:columns:` has been added to `RLMTransaction`.
+*  Added keyed subscripting for RLMTable's first column if column is of type RLMTypeString.
+*  `setRow:atIndex:` has been added to `RLMTable`
 
 -------------